+++ src/simulation/elements/LIFE.cpp
@@ -80,9 +80,9 @@ static int graphics(GRAPHICS_FUNC_ARGS)
 		else
 		{
 			auto mul = (cpart->tmp2 - 1) / float(states - 2);
-			*colr = PIXR(colour1) * mul + PIXR(colour2) * (1.f - mul);
-			*colg = PIXG(colour1) * mul + PIXG(colour2) * (1.f - mul);
-			*colb = PIXB(colour1) * mul + PIXB(colour2) * (1.f - mul);
+			*colr = int(PIXR(colour1) * mul + PIXR(colour2) * (1.f - mul));
+			*colg = int(PIXG(colour1) * mul + PIXG(colour2) * (1.f - mul));
+			*colb = int(PIXB(colour1) * mul + PIXB(colour2) * (1.f - mul));
 		}
 	}
 	*pixel_mode |= NO_DECO;
